Fixed stored data size incorrect issues, due to bad Dataflash block erasing algorithm. Cleanup, ready for V3.0 BETA testing.

+++ Dataflash.h
@@ -50,7 +50,7 @@
 	
 	// PROTOTYPES:
 	uint8_t DF_CheckCorrectOnboardChip(void) ATTR_WARN_UNUSED_RESULT;
-	void    DF_EraseBlock(const uint16_t BlockToErase);
+	void    DF_EraseBlock(const uint8_t BlockToErase);
 	void    DF_CopyPage(const uint16_t PageAddress, uint8_t Operation);
 	void    DF_ContinuousReadEnable(const uint16_t PageAddress, const uint16_t BuffAddress);
 	void    DF_BufferWriteEnable(const uint16_t BuffAddress);
